Added menu to `InputButton.Menu` and keybinding to the `QuickShop`
