+++ code/lang/TTTLanguage.cs
@@ -81,3 +81,35 @@ namespace TTTReborn.Globalization
         }
     }
 }
+
+namespace TTTReborn.Player
+{
+    using Globalization;
+
+    public partial class TTTPlayer
+    {
+        [ClientCmd("ttt_language")]
+        public static void ChangeLanguage(string name = null)
+        {
+            if (name is null)
+            {
+                Log.Info($"Your current language is set to '{TTTLanguage.ActiveLanguage.Data.Name}' ('{TTTLanguage.ActiveLanguage.Data.Code}').");
+
+                return;
+            }
+
+            Language language = TTTLanguage.GetLanguageByCode(name);
+
+            if (language is null)
+            {
+                Log.Warning($"Language '{name}' does not exist. Please enter an ISO (tag) code (http://www.lingoes.net/en/translator/langcode.htm).");
+
+                return;
+            }
+
+            Log.Warning($"You set your language to '{language.Data.Name}'.");
+
+            Settings.ClientSettings.Instance.Language = language.Data.Code;
+        }
+    }
+}
