+++ code/player/TTTPlayer.Commands.Client.cs
@@ -20,35 +20,5 @@ namespace TTTReborn.Player
                 Log.Info($"Player (ID: '{i}'): {playerList[i].Name}");
             }
         }
-
-        [ClientCmd(Name = "ttt_menu")]
-        public static void ToggleMenu()
-        {
-            Menu.Instance.IsShowing = !Menu.Instance.IsShowing;
-        }
-
-        [ClientCmd("ttt_language")]
-        public static void ChangeLanguage(string name = null)
-        {
-            if (name is null)
-            {
-                Log.Info($"Your current language is set to '{TTTLanguage.ActiveLanguage.Data.Name}' ('{TTTLanguage.ActiveLanguage.Data.Code}').");
-
-                return;
-            }
-
-            Language language = TTTLanguage.GetLanguageByCode(name);
-
-            if (language is null)
-            {
-                Log.Warning($"Language '{name}' does not exist. Please enter an ISO (tag) code (http://www.lingoes.net/en/translator/langcode.htm).");
-
-                return;
-            }
-
-            Log.Warning($"You set your language to '{language.Data.Name}'.");
-
-            Settings.ClientSettings.Instance.Language = language.Data.Code;
-        }
     }
 }
